deployment/requirements.txt
aws-cdk-lib==2.62.2
constructs>=10.0.0,<11.0.0
sagemaker
